refactor: tag-log-v1 & v2 -> tag-log
- #189
