--- src/tag-log-v2/dto/inout.dto.ts
@@ -1,6 +1,0 @@
-import InOut from 'src/enums/inout.enum';
-
-export class InOutDto {
-  log: Date | null;
-  inout: InOut;
-}
