--- src/tag-log-v2/dto/pair-info.dto.ts
@@ -1,4 +1,0 @@
-export class PairInfoDto {
-  in_device: number;
-  out_device: number;
-}
