+++ src/tag-log/dto/user-accumulation.type.v1.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 
-export class UserAccumulationType {
+export class UserAccumulationTypeV1 {
   @ApiProperty({
     description: '일별 누적시간',
     example: 12345,
