+++ src/tag-log/repository/interface/pair-info-repository.interface.ts
@@ -1,4 +1,4 @@
-import { PairInfoDto } from 'src/tag-log-v1/dto/pair-info.dto';
+import { PairInfoDto } from 'src/tag-log/dto/pair-info.dto';
 
 export interface IPairInfoRepository {
   /**
