+++ src/tag-log/repository/mysql/pair-info.repository.ts
@@ -1,8 +1,8 @@
 import { InjectRepository } from '@nestjs/typeorm';
-import { PairInfoDto } from 'src/tag-log-v1/dto/pair-info.dto';
-import { IPairInfoRepository } from '../interface/pair-info-repository.interface';
-import { Repository } from 'typeorm';
 import { PairInfo } from 'src/entities/pair-info.entity';
+import { PairInfoDto } from 'src/tag-log/dto/pair-info.dto';
+import { Repository } from 'typeorm';
+import { IPairInfoRepository } from '../interface/pair-info-repository.interface';
 
 export class PairInfoRepository implements IPairInfoRepository {
   constructor(
